[docs] remarks for the signature of `Ping` class
The title for the section of the `Ping` class was updated to 'Abstract
Class', with the intention of noting that the class is not usable unless
extended.
All the links to the section were updated and some code blocks where
labeled with the language (`cmd bash`, `text`).
A simple addition was added to the clause "Ping is *an* EventEmitter"